[feat]: add validation for password, meetings endtime, task deadline. add author to task

+++ app/presentation/companies.py
@@ -2,11 +2,12 @@ from app.infrastructure.database.session import get_session
 from app.infrastructure.schemas.companies import CompanyCreate, Company
 from app.infrastructure.schemas.users import User
 from app.infrastructure.services.companies import CompanyService
-from app.core.permissions import user_admin_permission
+from app.core.permissions import user_admin_permission, user_manager_permission
 from fastapi import APIRouter, Depends
 from fastapi import HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
 from uuid import UUID
+
 
 companies = APIRouter(
     prefix="/companies",
@@ -55,7 +56,7 @@ async def get_company(
 async def add_user_to_company(
     user_id: UUID,
     company_id: UUID,
-    current_user: User = Depends(user_admin_permission),
+    current_user: User = Depends(user_manager_permission),
     session: AsyncSession = Depends(get_session),
 ):
     try:
@@ -74,10 +75,12 @@ async def add_user_to_company(
 async def delete_user_to_company(
     user_id: UUID,
     company_id: UUID,
-    current_user: User = Depends(user_admin_permission),
+    manager_permission: User = Depends(user_manager_permission),
     session: AsyncSession = Depends(get_session),
 ):
     try:
+        if not manager_permission.company_id == company_id:
+            raise ValueError("Данный менеджер не привязан к этой компании")
         company_response = await CompanyService(session).delete_user_from_company(
             user_id=user_id, company_id=company_id
         )
@@ -85,3 +88,5 @@ async def delete_user_to_company(
         return company
     except Exception as _e:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(_e))
+    except ValueError as _ve:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=str(_ve))
